+++ toshi_hazard_store/config.py
@@ -15,7 +15,7 @@ def boolean_env(environ_name: str, default: str = 'FALSE') -> bool:
 IS_OFFLINE = boolean_env(
     'SLS_OFFLINE'
 )  # set by serverless-wsgi plugin, and used only when THS is included in a WSGI test
-REGION = os.getenv('NZSHM22_HAZARD_STORE_REGION')
+REGION = os.getenv('NZSHM22_HAZARD_STORE_REGION', "us-east-1")
 DEPLOYMENT_STAGE = os.getenv('NZSHM22_HAZARD_STORE_STAGE', 'LOCAL').upper()
 NUM_BATCH_WORKERS = int(os.getenv('NZSHM22_HAZARD_STORE_NUM_WORKERS', 1))
 LOCAL_CACHE_FOLDER = os.getenv('NZSHM22_HAZARD_STORE_LOCAL_CACHE')